+++ NEWS.md
@@ -1,6 +1,11 @@
 
 # lifecycle 0.1.0.9000
 
+* New `expect_deprecated()` and `expect_defunct()` functions for
+  testting lifecycle warnings and errors. `expect_deprecated()`
+  automatically sets the `lifecycle_verbosity` option to `"warning"`
+  to enforce warnings at each invokation rather than once per session.
+
 * New syntax `"foo(arg = 'can\\'t be a baz')"` to describe that specific inputs
   for an argument are deprecated (#30, @krlmlr).